not going to do InterMineCoord; revert some things back to Integers

+++ intermine/objectstore/main/src/org/intermine/objectstore/query/Clob.java
@@ -76,6 +76,6 @@ public class Clob implements QuerySelectable
      */
     @Override
     public int hashCode() {
-        return clobId;
+        return clobId.intValue();
     }
 }
